feature LocationService를 추상화하고 테스트용 DemoLocationService 구현
